requirements.txt
python-binance==1.0.19
click==8.1.3
rich==13.3.2
pyyaml==6.0
tabulate==0.9.0
python-dotenv==1.0.0